rename hooks/useCustomSounds -> hooks/sounds
